Removed long long statement

+++ holberton.h
@@ -13,7 +13,7 @@ int _printf(const char *format, ...);
 void _print_rev_recursion(char *s);
 void rot13(char *s);
 void convert_to_hex(char *s);
-void print_number(unsigned long long int n);
+void print_number(unsigned long int n);
 void convert_to_binary(unsigned int s);
 /*
  *void print_to_console(char *str);
